+++ src/main/kotlin/org/powernukkit/converters/conversion/adapter/BlockStateAdapter.kt
@@ -1,0 +1,55 @@
+/*
+ * PowerNukkit Universal Worlds & Converters for Minecraft
+ * Copyright (C) 2020  José Roberto de Araújo Júnior
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.powernukkit.converters.conversion.adapter
+
+import org.powernukkit.converters.conversion.context.IncompleteBlockState
+import org.powernukkit.converters.platform.api.BlockContainer
+import org.powernukkit.converters.platform.api.Platform
+import org.powernukkit.converters.platform.api.block.PlatformBlock
+import org.powernukkit.converters.platform.api.block.PlatformBlockState
+
+/**
+ * @author joserobjr
+ * @since 2020-10-18
+ */
+interface BlockStateAdapter<FromPlatform : Platform<FromPlatform>, ToPlatform : Platform<ToPlatform>> {
+    fun adaptBlockState(
+        fromPlatform: FromPlatform,
+        toPlatform: ToPlatform,
+        fromState: PlatformBlockState<FromPlatform>,
+        incomplete: IncompleteBlockState<ToPlatform>,
+        current: List<PlatformBlockState<ToPlatform>>?,
+        fromLayer: Int,
+        fromLayers: List<PlatformBlockState<FromPlatform>>,
+        fromBlock: PlatformBlock<FromPlatform>,
+        fromContainer: BlockContainer<FromPlatform>
+    ): List<PlatformBlockState<ToPlatform>>?
+
+    fun adaptBlockStateList(
+        fromPlatform: FromPlatform,
+        toPlatform: ToPlatform,
+        fromState: PlatformBlockState<FromPlatform>,
+        incomplete: IncompleteBlockState<ToPlatform>,
+        current: List<PlatformBlockState<ToPlatform>>?,
+        fromLayer: Int,
+        fromLayers: List<PlatformBlockState<FromPlatform>>,
+        fromBlock: PlatformBlock<FromPlatform>,
+        fromContainer: BlockContainer<FromPlatform>
+    ): List<PlatformBlockState<ToPlatform>>?
+}
