+++ src/main/kotlin/org/powernukkit/converters/conversion/context/BlockLayerConversionContext.kt
@@ -16,28 +16,20 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.powernukkit.converters.converter
+package org.powernukkit.converters.conversion.context
 
-import org.powernukkit.converters.platform.api.BlockContainer
 import org.powernukkit.converters.platform.api.Platform
-import org.powernukkit.converters.platform.api.block.PlatformBlock
 import org.powernukkit.converters.platform.api.block.PlatformBlockState
-import org.powernukkit.converters.platform.api.block.PlatformBlockType
 
 /**
  * @author joserobjr
- * @since 2020-10-17
+ * @since 2020-10-18
  */
-interface BlockTypeAdapter<FromPlatform : Platform<FromPlatform>, ToPlatform : Platform<ToPlatform>> {
-    fun adaptBlockType(
-        fromType: PlatformBlockType<FromPlatform>,
-        fromState: PlatformBlockState<FromPlatform>,
-        fromLayer: Int,
-        fromLayers: List<PlatformBlockState<FromPlatform>>,
-        fromBlock: PlatformBlock<FromPlatform>,
-        fromContainer: BlockContainer<FromPlatform>,
-        fromPlatform: FromPlatform,
-        toPlatform: ToPlatform,
-        currentResult: PlatformBlockType<ToPlatform>?
-    ): PlatformBlockType<ToPlatform>?
+data class BlockLayerConversionContext<FromPlatform : Platform<FromPlatform>, ToPlatform : Platform<ToPlatform>>(
+    val layersContext: FullBlockLayersConversionContext<FromPlatform, ToPlatform>,
+    val fromLayer: Int,
+    val fromBlockState: PlatformBlockState<FromPlatform>,
+) {
+    var requiresAdapter: Boolean = false
+    var toBlockStateLayers: List<PlatformBlockState<ToPlatform>>? = null
 }
