+++ src/main/kotlin/org/powernukkit/converters/conversion/context/BlockPropertyValuesConverter.kt
@@ -16,8 +16,10 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.powernukkit.converters.converter
+package org.powernukkit.converters.conversion.context
 
+import org.powernukkit.converters.conversion.adapter.Adapters
+import org.powernukkit.converters.conversion.adapter.BlockPropertyValuesAdapter
 import org.powernukkit.converters.platform.api.BlockContainer
 import org.powernukkit.converters.platform.api.NamespacedId
 import org.powernukkit.converters.platform.api.Platform
