--- src/main/kotlin/org/powernukkit/converters/converter/Adapters.kt
@@ -1,31 +1,0 @@
-/*
- * PowerNukkit Universal Worlds & Converters for Minecraft
- * Copyright (C) 2020  José Roberto de Araújo Júnior
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package org.powernukkit.converters.converter
-
-/**
- * @author joserobjr
- * @since 2020-10-18
- */
-data class Adapters<I, A>(
-    val firstAdapters: List<A> = emptyList(),
-    val fromAdapters: Map<I, List<A>> = emptyMap(),
-    val toAdapters: Map<I, List<A>> = emptyMap(),
-    val lastAdapters: List<A> = emptyList(),
-    val lastToAdapters: Map<I, List<A>> = emptyMap(),
-)
