--- src/main/kotlin/org/powernukkit/converters/converter/EntityConverter.kt
@@ -1,45 +1,0 @@
-/*
- * PowerNukkit Universal Worlds & Converters for Minecraft
- * Copyright (C) 2020  José Roberto de Araújo Júnior
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package org.powernukkit.converters.converter
-
-import org.powernukkit.converters.platform.api.BlockContainer
-import org.powernukkit.converters.platform.api.Platform
-import org.powernukkit.converters.platform.api.block.PlatformBlock
-import org.powernukkit.converters.platform.api.block.PlatformBlockEntity
-import org.powernukkit.converters.platform.api.block.PlatformBlockState
-import org.powernukkit.converters.platform.api.entity.PlatformEntity
-
-/**
- * @author joserobjr
- * @since 2020-10-17
- */
-open class EntityConverter<FromPlatform : Platform<FromPlatform>, ToPlatform : Platform<ToPlatform>>(
-    val fromPlatform: FromPlatform,
-    val toPlatform: ToPlatform,
-) {
-    open fun convert(
-        fromBlock: PlatformBlock<FromPlatform>,
-        fromContainer: BlockContainer<FromPlatform>,
-        convertedLayers: List<PlatformBlockState<ToPlatform>>,
-        convertedBlockEntity: PlatformBlockEntity<ToPlatform>?
-    ): List<PlatformEntity<ToPlatform>> {
-        TODO("Not yet implemented")
-    }
-
-}
